move the framebuffer module to the common crate
