+++ common/src/framebuffer/framebuffer.rs
