+++ common/src/lib.rs
@@ -2,3 +2,6 @@
 
 #[path="font/font.rs"]
 pub mod font;
+
+#[path="framebuffer/framebuffer.rs"]
+pub mod framebuffer;
